fix backuping

+++ initdb.sql
@@ -1,7 +1,7 @@
 BEGIN TRANSACTION;
 CREATE TABLE IF NOT EXISTS `sync` (
 	`id`	INTEGER PRIMARY KEY AUTOINCREMENT,
-	`date`	NUMERIC,
+	`date`	REAL,
 	`source`	TEXT,
 	`zip_name`	TEXT,
 	`zip_pass_salt`  TEXT
